Refactor apps. Wrap setup and teardown environment.
