+++ tests/test_middleware.py
@@ -1,12 +1,14 @@
 import os
 from io import StringIO
 from os import path
-from unittest import TestSuite, TextTestRunner, mock
+from unittest import TestLoader, TextTestRunner, mock
 from unittest.mock import MagicMock
 
 from django.core.exceptions import MiddlewareNotUsed
 from django.test import TestCase, override_settings
+from django.test.runner import DiscoverRunner
 from test_query_counter.apps import RequestQueryCountConfig
+from test_query_counter.manager import RequestQueryCountManager
 from test_query_counter.middleware import Middleware
 
 
@@ -19,7 +21,8 @@ class TestMiddleWare(TestCase):
                 kwargs['stream'] = StringIO()
                 super().__init__(*args, **kwargs)
 
-        self.test_runner = StringIOTextRunner()
+        self.test_runner = DiscoverRunner()
+        self.test_runner.test_runner = StringIOTextRunner
 
     def tearDown(self):
         try:
@@ -42,12 +45,12 @@ class TestMiddleWare(TestCase):
             def test_foo(self):
                 self.client.get('/url-1')
 
-        suite = TestSuite()
-        suite.addTest(Test('test_foo'))
+        self.test_runner.setup_test_environment()
+        self.test_runner.run_suite(TestLoader().loadTestsFromTestCase(
+            testCaseClass=Test))
+        self.test_runner.teardown_test_environment()
 
-        result = self.test_runner.run(suite)
-
-        self.assertEqual(result.queries.total, 1)
+        self.assertEqual(RequestQueryCountManager.queries.total, 1)
 
     def test_case_injected_two_tests(self):
         class Test(TestCase):
@@ -57,13 +60,11 @@ class TestMiddleWare(TestCase):
             def test_bar(self):
                 self.client.get('/url-2')
 
-        suite = TestSuite()
-        suite.addTest(Test('test_foo'))
-        suite.addTest(Test('test_bar'))
+        self.test_runner.run_suite(
+            TestLoader().loadTestsFromTestCase(testCaseClass=Test)
+        )
 
-        result = self.test_runner.run(suite)
-
-        self.assertEqual(result.queries.total, 2)
+        self.assertEqual(RequestQueryCountManager.queries.total, 2)
 
     @override_settings(TEST_QUERY_COUNTER={'ENABLE': False})
     def test_case_disable_setting(self):
@@ -74,12 +75,11 @@ class TestMiddleWare(TestCase):
             def test_bar(self):
                 self.client.get('/url-2')
 
-        suite = TestSuite()
-        suite.addTest(Test('test_foo'))
-        suite.addTest(Test('test_bar'))
-
-        result = self.test_runner.run(suite)
-        self.assertFalse(hasattr(result, 'queries'))
+        self.test_runner.run_tests(
+            None,
+            TestLoader().loadTestsFromTestCase(testCaseClass=Test)
+        )
+        self.assertIsNone(RequestQueryCountManager.queries)
 
     @override_settings(TEST_QUERY_COUNTER={'ENABLE': False})
     def test_disabled(self):
@@ -92,13 +92,13 @@ class TestMiddleWare(TestCase):
             def test_foo(self):
                 self.client.get('/url-1')
 
-        suite = TestSuite()
-        suite.addTest(Test('test_foo'))
-
         self.assertFalse(path.exists(
             RequestQueryCountConfig.get_setting('DETAIL_PATH'))
         )
-        self.test_runner.run(suite)
+        self.test_runner.run_tests(
+            None,
+            TestLoader().loadTestsFromTestCase(testCaseClass=Test)
+        )
         self.assertTrue(path.exists(
             RequestQueryCountConfig.get_setting('DETAIL_PATH'))
         )
